feat: mechanical extraction of completions and code_actions
